Add files via upload
This repo contains 3 python files: McCabe-Thiele.py which was used in my job to estimate the number of plates in the column during the rectification process. The function was checked benzene data. The other 2 are pxy.py and txy.py which are the functions that allow us to create used phase diagrams widely used in Chemical Engineering to determine the composition of the mixture. These 3 were my first "bigger" projects in Python.